HWKBTM-82 Refactor to share common test capabilities across integration tests
